--- .env
@@ -1,3 +1,0 @@
- PORT=5000
-MONGO_URI=mongodb://127.0.0.1:27017/doctorBookingDB
-JWT_SECRET=yourSecretKey
