--- adminController.js
@@ -1,19 +1,0 @@
-const Admin = require("../models/Admin");
-const jwt = require("jsonwebtoken");
-
-// Admin login
-exports.loginAdmin = async (req, res) => {
-  try {
-    const { username, password } = req.body;
-    const admin = await Admin.findOne({ username });
-    if (!admin) return res.status(404).json({ message: "Admin not found" });
-
-    const isMatch = await admin.comparePassword(password);
-    if (!isMatch) return res.status(400).json({ message: "Invalid password" });
-
-    const token = jwt.sign({ id: admin._id, role: "admin" }, process.env.JWT_SECRET, { expiresIn: "1d" });
-    res.json({ token, message: "Admin login successful" });
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-};
