--- server.js
@@ -1,37 +1,0 @@
-const express = require("express");
-const mongoose = require("mongoose");
-const bodyParser = require("body-parser");
-const dotenv = require("dotenv");
-
-dotenv.config();
-const app = express();
-const PORT = process.env.PORT || 5000;
-
-// Middleware
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
-
-// Import routes
-const userRoutes = require("./routes/userRoutes");
-const doctorRoutes = require("./routes/doctorRoutes");
-const adminRoutes = require("./routes/adminRoutes");
-
-// Use routes
-app.use("/api/users", userRoutes);
-app.use("/api/doctors", doctorRoutes);
-app.use("/api/admins", adminRoutes);
-
-// Test route
-app.get("/", (req, res) => {
-  res.send("Doctor Booking Backend Running!");
-});
-
-// Connect MongoDB
-mongoose.connect(process.env.MONGO_URI)
-  .then(() => console.log("MongoDB connected"))
-  .catch(err => console.log("MongoDB connection error:", err));
-
-// Start server
-app.listen(PORT, () => {
-  console.log("Server running on http://localhost:${PORT}");
-});
