+++ server/middleware/adminAuth.js
@@ -1,0 +1,18 @@
+const jwt = require("jsonwebtoken");
+
+const adminAuth = (req, res, next) => {
+  try {
+    const token = req.headers.authorization?.split(" ")[1];
+    if (!token) return res.status(401).json({ message: "No token provided" });
+
+    const decoded = jwt.verify(token, process.env.JWT_SECRET);
+    if (decoded.role !== "admin") return res.status(403).json({ message: "Access denied. Not admin." });
+
+    req.admin = decoded;
+    next();
+  } catch (err) {
+    res.status(400).json({ message: "Invalid token" });
+  }
+};
+
+module.exports = adminAuth;
