--- userController.js
@@ -1,45 +1,0 @@
-const User = require("../models/User");
-const jwt = require("jsonwebtoken");
-
-// Register a new user
-exports.registerUser = async (req, res) => {
-  try {
-    const { name, email, password, phone } = req.body;
-    let user = await User.findOne({ email });
-    if (user) return res.status(400).json({ message: "User already exists" });
-
-    user = new User({ name, email, password, phone });
-    await user.save();
-
-    res.status(201).json({ message: "User registered successfully" });
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-};
-
-// User login
-exports.loginUser = async (req, res) => {
-  try {
-    const { email, password } = req.body;
-    const user = await User.findOne({ email });
-    if (!user) return res.status(404).json({ message: "User not found" });
-
-    const isMatch = await user.comparePassword(password);
-    if (!isMatch) return res.status(400).json({ message: "Invalid password" });
-
-    const token = jwt.sign({ id: user._id }, process.env.JWT_SECRET, { expiresIn: "1d" });
-    res.json({ token, message: "Login successful" });
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-};
-
-// Get all users
-exports.getUsers = async (req, res) => {
-  try {
-    const users = await User.find().select("-password");
-    res.json(users);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-};
